requirements.txt
Flask>=2.3
Werkzeug>=2.3
tensorflow>=2.12
numpy>=1.24
Pillow>=10.0
gTTS>=2.5
python-dotenv>=1.0
gunicorn>=21.2
